+++ src/components/complex/object-holdings/deals/deal-item/floors/head/DealFloorHead.scss
@@ -14,6 +14,16 @@
     }
   }
 
+  &-button {
+    background: none;
+    border: none;
+    cursor: pointer;
+    &:hover {
+      opacity: 0.6;
+    }
+
+  }
+
   &-checkbox {
     border-radius: 50%;
     position: relative;
