+++ src/components/complex/object-holdings/forms/plot-create-form/PlotCreateForm.vue
@@ -531,17 +531,101 @@
 import {defineComponent} from "vue";
 import {helpers, required, minValue} from "@vuelidate/validators";
 import {ObjectForms} from "@/components/complex/object-holdings/forms/mixin";
+import {mapActions} from "vuex";
+import {yandexmap} from "@/utils";
+import {
+	entryTerritoryTypes,
+	feeTypes,
+	gasTypes,
+	guardTypes,
+	internetTypes,
+	landCategoryTypes,
+	landscapeTypes,
+	ownTypesLand,
+	waterTypes
+} from "@/const/constTypes";
 
 
 export default defineComponent({
 	name: "PlotCreateForm",
 	mixins: [ObjectForms],
 	data() {
-		return {}
+		return {
+			form: {
+				address: null,
+				company_id: null,
+				object_type: [],
+				purposes: [],
+				test_only: null,
+				area_field_full: null,
+				land_length: null,
+				land_width: null,
+				own_type_land: null,
+				land_category: null,
+				landscape_type: null,
+				cadastral_number_land: null,
+				first_line: null,
+				field_allow_usage: null,
+				power: null,
+				power_value: null,
+				sewage_central: null,
+				sewage_central_value: null,
+				sewage_rain: null,
+				water: null,
+				water_value: null,
+				water_type: null,
+				gas: null,
+				gas_value: null,
+				gas_type: null,
+				steam: null,
+				steam_value: null,
+				phone_line: null,
+				internet: null,
+				internet_type: null,
+				guard: null,
+				guard_type: null,
+				video_control: null,
+				access_control: null,
+				security_alert: null,
+				barrier: null,
+				fence_around_perimeter: null,
+				railway: null,
+				railway_value: null,
+				entry_territory: null,
+				entry_territory_type: null,
+				entrance_type: null,
+				parking_car: null,
+				parking_car_value: null,
+				parking_lorry: null,
+				parking_lorry_value: null,
+				parking_truck: null,
+				parking_truck_value: null,
+				canteen: null,
+				hostel: null,
+				description: null,
+				layouts: [],
+				layoutsList: [],
+				panoramas: [],
+				panoramasList: [],
+				presentations: [],
+				presentationsList: [],
+				ownerShipDocuments: [],
+				ownerShipDocumentsList: [],
+				photos: [],
+				photosList: [],
+
+			},
+		}
 	},
 	validations() {
 		return {
 			form: {
+				address: {
+					required: helpers.withMessage("заполните поле", required),
+				},
+				company_id: {
+					required: helpers.withMessage("заполните поле", required),
+				},
 				cadastral_number_land: {
 					required: helpers.withMessage("заполните поле", required),
 				},
@@ -699,34 +783,58 @@ export default defineComponent({
 			}
 		}
 	},
-	methods: {},
+	methods: {
+		...mapActions([
+			"SEARCH_COMPANIES",
+		]),
+		onChangeAddress() {
+			console.log("CHANGE ADDRESS")
+		},
+		async onChangeCompany() {
+			console.log("CHANGE COMPANY")
+		},
+		async searchCompany() {
+			return ["ОАО Тореадор", "ЗАО ИнкЛомМет"]
+		},
+		async searchAddress(query) {
+			return await yandexmap.getAddress(query);
+		},
+		selectObjectType(isSelected, type) {
+			this.form.object_type = this.form.object_type.filter(
+					(item) => item !== type
+			);
+			if (isSelected) {
+				this.form.object_type.push(type);
+			}
+		},
+	},
 	computed: {
 		ownTypeLandOptions() {
-			return Object.values(this.ownTypesLand);
+			return Object.values(ownTypesLand);
 		},
 		landCategoryOptions() {
-			return Object.values(this.landCategoryTypes);
+			return Object.values(landCategoryTypes);
 		},
 		landscapeTypeOptions() {
-			return Object.values(this.landscapeTypes);
+			return Object.values(landscapeTypes);
 		},
 		waterTypeOptions() {
-			return Object.values(this.waterTypes);
+			return Object.values(waterTypes);
 		},
 		gasTypeOptions() {
-			return Object.values(this.gasTypes);
+			return Object.values(gasTypes);
 		},
 		internetTypeOptions() {
-			return Object.values(this.internetTypes);
+			return Object.values(internetTypes);
 		},
 		guardTypeOptions() {
-			return Object.values(this.guardTypes)
+			return Object.values(guardTypes)
 		},
 		feeTypeOptions() {
-			return Object.values(this.feeTypes)
+			return Object.values(feeTypes)
 		},
 		entryTerritoryTypeOptions() {
-			return Object.values(this.entryTerritoryTypes)
+			return Object.values(entryTerritoryTypes)
 		}
 	}
 })
